fix(shaderlib): support fractional libretro scales
Changelog: updated
Guide: n/a: no guide changes
Known-Discrepancies: n/a: no discrepancy changes
S3K-Known-Discrepancies: n/a: no S3K discrepancy changes
Agent-Docs: n/a: no agent doc changes
Configuration-Docs: n/a: no configuration doc changes
Skills: n/a: no skill changes

+++ src/main/java/com/openggf/graphics/shaderlib/DisplayShaderPass.java
@@ -4,7 +4,7 @@ public record DisplayShaderPass(
         String vertexSource,
         String fragmentSource,
         GlslShape shape,
-        int scale,
+        double scale,
         ScaleType scaleType,
         boolean filterLinear,
         WrapMode wrapMode) {
